requirements.txt
absl-py==2.1.0
actionlib==1.14.0
agentlace @ git+https://github.com/youliangtan/agentlace.git@cf2c337c5e3694cdbfc14831b239bd657bc4894d
angles==1.9.13
annotated-types==0.7.0
array_record==0.6.0
astunparse==1.6.3
blinker==1.9.0
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
chex==0.1.88
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distlib==0.3.9
distrax==0.1.5
distro==1.9.0
dm-env==1.6
dm-tree==0.1.8
dm_robotics-transformations==0.8.1
docker-pycreds==0.4.0
docstring_parser==0.16
docutils==0.21.2
dynamic-reconfigure==1.7.3
easyhid==0.0.10
einops==0.8.0
etils==1.11.0
evdev==1.7.1
Farama-Notifications==0.0.4
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
flax==0.10.2
fonttools==4.55.3
fsspec==2024.12.0
gast==0.6.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gitdb==4.0.11
GitPython==3.1.43
glfw==2.8.0
google-pasta==0.2.0
grpcio==1.68.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.12.1
hidapi==0.14.0.post4
humanize==4.11.0
identify==2.6.3
idna==3.10
image-geometry==1.16.2
imageio==2.36.1
imageio-ffmpeg==0.5.1
immutabledict==4.2.1
importlib_resources==6.4.5
interactive-markers==1.12.0
itsdangerous==2.2.0
jax==0.4.35
jax-cuda12-pjrt==0.4.35
jax-cuda12-plugin==0.4.35
jaxlib==0.4.34
Jinja2==3.1.5
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
keras==3.7.0
kiwisolver==1.4.7
laser_geometry==1.6.7
libclang==18.1.1
lxml==5.3.0
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
message-filters==1.17.0
ml-dtypes==0.4.1
ml_collections==1.0.0
moviepy==2.1.1
msgpack==1.1.0
mujoco==2.3.7
namex==0.0.8
natsort==8.4.0
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==1.26.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optax==0.2.4
optree==0.13.1
orbax-checkpoint==0.10.3
packaging==24.2
pillow==10.4.0
platformdirs==4.3.6
pre-commit==3.3.3
proglog==0.1.10
promise==2.3
protobuf==3.20.3
psutil==6.1.1
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pymodbus==2.5.3
pynput==1.7.7
PyOpenGL==3.1.7
pyparsing==3.2.0
pyquaternion==0.9.9
pyrealsense2==2.55.1.6486
pyserial==3.5
pyspacemouse==1.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-qt-binding==0.4.4
python-xlib==0.33
PyYAML==6.0.2
pyzmq==26.2.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.32.3
resource_retriever==1.12.7
rich==13.9.4
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospkg==1.5.1
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rviz==1.14.25
scipy==1.11.4
sensor-msgs==1.13.1
sentry-sdk==2.19.2
-e ./serl_robot_infra
setproctitle==1.3.4
simple-parsing==0.1.6
simplejson==3.19.3
six==1.17.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
smmap==5.0.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-datasets==4.9.7
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensorflow-probability==0.25.0
tensorstore==0.1.71
termcolor==2.5.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
tf_keras==2.18.0
toml==0.10.2
toolz==1.0.0
topic-tools==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.28.0
wandb==0.19.1
Werkzeug==3.1.3
wrapt==1.17.0
xacro==1.14.18
zipp==3.21.0
zmq==0.0.0



pyrealsense2==2.56.5.9235
pyquaternion==0.9.9
cffi==2.0.0
easyhid==0.0.10
pycparser==2.23
aiofiles==25.1.0
etils==1.13.0
flax==0.10.7
humanize==4.14.0
importlib_resources==6.5.2
msgpack==1.1.2
nest_asyncio==1.6.0
orbax-checkpoint==0.11.30
simplejson==3.20.2
tensorstore==0.1.78
treescope==0.1.10
zipp==3.23.0
flask
ur_rtde
requests
